Change of method, Added an escape mechanism
Can now escape from battles and also changed the method used to change
attack modifier form attackDown to attackChange. Usable actions added to
battle menu.

+++ RPGProject/Battle.java
@@ -14,7 +14,9 @@ public class Battle {
 		turnCount = 0;
 		Scanner s = new Scanner(System.in);
 		while (!battlePlay.isDead() && !battleEnem.isDead()) {
+			battleActions();
 			action = s.next();
+			System.out.println("Turn #" + turnCount);
 			if (battlePlay.isInflicted()) {
 				for (Ailment a : battlePlay.ailments()) {
 					a.affect(battlePlay);
@@ -25,14 +27,23 @@ public class Battle {
 					a.affect(battleEnem);
 				}
 			}
-			if (battleEnem.getSpd() >= battlePlay.getSpd()) {
-				battleEnem.attack(battlePlay);
-				battlePlay.attack(battleEnem, action);
+			if (action.equals("a") || action.equals("A") || action.equals("k") || action.equals("K")){
+				if (battleEnem.getSpd() >= battlePlay.getSpd()) {
+					battleEnem.attack(battlePlay);
+					battlePlay.attack(battleEnem, action);
+				} else {
+					battlePlay.attack(battleEnem, action);
+					battleEnem.attack(battlePlay);
+				}
 			} else {
 				battlePlay.attack(battleEnem, action);
-				battleEnem.attack(battlePlay);
+				if (battlePlay.canRunAway()) {
+					endBattle();
+				} else {
+					battleEnem.attack(battlePlay);
+				}
 			}
-			System.out.println("Turn #" + turnCount);
+			
 			System.out.println(battlePlay.getHP());
 			System.out.println(battleEnem.getHP());
 			System.out.println();
@@ -42,13 +53,28 @@ public class Battle {
 			System.out.println("Game Over");
 			System.exit(0);
 		} else {
-			battlePlay.getInventory().add(battleEnem.drop());
-			battlePlay.gainEXP(battleEnem.dropEXP());
-			battlePlay.levelUp();
+			if (battlePlay.canRunAway()) {
+				System.out.println("You leave the enemy behind!");
+				battlePlay.setRunAway(false);
+			} else {
+				battlePlay.getInventory().add(battleEnem.drop());
+				battlePlay.gainEXP(battleEnem.dropEXP());
+				battlePlay.levelUp();
+			}
 		}
 	}
 
 	public void battleActions() {
-		System.out.println();
+		System.out.println("----------Player Actions----------");
+		System.out.println("[A]ttack\t[I]nventory");
+		System.out.println("S[k]ills\t[S]kip");
+		System.out.println("[R]un   \t[D]efend");
+		System.out.print("Please choose an action: ");
+	}
+
+	public void endBattle() {
+		System.out.println("You ran away!");
+		battleEnem.setHP(-battleEnem.getMaxHP());
+		//battleEnem.setAtt(-battleEnem.getAtt());
 	}
 }
